Устранение замечания: все вьюхи рекомендую делать приватными, для обновления их стейта нужно использовать публичные методы
